+++ js/utils.js
@@ -1,0 +1,21 @@
+function getRandomPositiveInteger (a, b) {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+}
+
+function getMaxLetter (currentLine, maxLine = 140) {
+  if (currentLine > maxLine) {
+    return false;
+  }
+  return true;
+}
+
+getMaxLetter(0, 10);
+
+export {getRandomPositiveInteger};
